+++ stories/components/card/wcs-card-content.stories.ts
@@ -1,0 +1,12 @@
+import { Meta, StoryFn, StoryObj } from '@storybook/web-components';
+import { html } from 'lit-html';
+
+const meta: Meta = {
+    title: 'Components/Card/Card Content',
+    component: 'wcs-card-content',
+};
+
+export default meta;
+
+const Template: StoryFn<Partial<any>> = () => html``;
+export const Default: StoryObj = Template.bind({});